Add GDAX real-time data listening
GDAXPipeline is required when loading real time data since it has to
connect to GDAX’s web socket. Loading historical data does not have to
be part of the object so the function has been separated from
GDAXPipeline.

A websocket library has been added to the project as a copy since this
project is not using a package management system. Another option was to
import the library through Eclipse’s properties but that would make the
project hard to maintain from multiple computers.